Replace isstructuresymmetric method
The original method seems to try to record if a structure symmetry pair already
has been found and therefore introduced the marker array `first`. For some corner cases
(easily to generate via sprand...) this fails. Attempts to fix this failed as well.
The new method is slightly less efficient (but uses no additional memory...),
but passes all tests. The performance penalty IMHO is be worth the correctness
and at least buys time for a more profound fix later on.