SparseCSCInterface sparspaklu! check sparsity pattern
sparsepaklu! now has a 'reuse_symbolic' option:
    sparspaklu!(lu::SparseSolver, m::SparseMatricCSC; reuse_symbolic=true)

If 'reuse_symbolic=true' (the default) then the sparsity pattern of 'm' is checked
and an error generated if it does not match that of the matrix used to create 'lu'

This matches the interface for the Julia SparseArrays stdlib UMFPACK lu!

Also some bug fixes:
    - sparspaklu! always mutate and update 'lu'
    - ErrorExceptions were created, but not actually thrown ! Fix by
    replacing ErrorException() with error()