Update sparspaklu! to use allow_pattern_change option
sparspaklu!(lu::SparseSolver, m::SparseMatrixCSC; allow_pattern_change=true)

If `allow_pattern_change = true` (the default) the sparse matrix `m` may have a nonzero pattern
different to that of the matrix used to create the LU factorization `lu`, in which case the ordering
and symbolic factorization are updated.

If `allow_pattern_change = false` an error is thrown if the nonzero pattern of `m` is different to that
of the matrix used to create the LU factorization `lu`.

If `lu` was created with option `factorize = false` then `lu` is always updated from `m` and `allow_pattern_change` is ignored.